feat add SupplierProductInfoComponent
