+++ packages/apps/stock/src/components/organisms/SupplierProductInfo/SupplierProductInfo.js
@@ -1,0 +1,68 @@
+/*
+ * Axelor Business Solutions
+ *
+ * Copyright (C) 2023 Axelor (<http://axelor.com>).
+ *
+ * This program is free software: you can redistribute it and/or  modify
+ * it under the terms of the GNU Affero General Public License, version 3,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import React from 'react';
+import {Text, Icon} from '@axelor/aos-mobile-ui';
+import {View, StyleSheet} from 'react-native';
+import {useSelector, useTranslator} from '@axelor/aos-mobile-core';
+
+const SupplierProductInfo = ({}) => {
+  const I18n = useTranslator();
+  const {supplierProductInfo} = useSelector(state => state.supplierCatalog);
+
+  return (
+    <>
+      {supplierProductInfo == null ||
+      Object.keys(supplierProductInfo).length === 0 ? null : (
+        <View style={styles.supplierInfoContainer}>
+          <Icon name="info-circle" size={20} />
+          <View style={styles.supplierInfo}>
+            <Text style={styles.text_important}>
+              {I18n.t('Stock_SupplierCatalog')}
+            </Text>
+            <Text>{`${I18n.t('Stock_Name')} : ${
+              supplierProductInfo?.productSupplierName
+            }`}</Text>
+            <Text>{`${I18n.t('Stock_Code')} : ${
+              supplierProductInfo?.productSupplierCode
+            }`}</Text>
+          </View>
+        </View>
+      )}
+    </>
+  );
+};
+
+const styles = StyleSheet.create({
+  supplierInfoContainer: {
+    marginVertical: '1%',
+    marginHorizontal: 16,
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+  supplierInfo: {
+    marginLeft: '3%',
+    flexDirection: 'column',
+  },
+  text_important: {
+    fontSize: 16,
+    fontWeight: 'bold',
+  },
+});
+
+export default SupplierProductInfo;
